Implement IFlightStatusProvider
Implement IFlightStatusProvider, which provides a general interface for classes
that provide repeated FlightState updates. FlightStatusController and TelloApi
now implement IFlightStatusProvider - TelloApi generates the state updates, and
is intended to send them to the FlightStatusController, which will then
propagate the information to the CDI. Currently, the TelloApi is wired directly
to the CDI, bypassing logging capabilities until they are implemented fully.