Convert solution to single exe project w/ modules
The Review and Flight modules (previously independently executable)
will be converted to DLLs and referenced by DroneNavigator.

Make the DroneNavigator project the entry point for the solution. DroneNavigator
will provide a landing page to select a mission to review (spawning the
MainWindow from the Review module) or to launch a new mission (spawning the
MainWindow from the Flight module).