Implement connect/disconnect button in Flight
Implements a button for connection and disconnection in the Flight module.
If the drone is not currently connected, this button attempts to start a
connection. If that connection attempt fails, an InfoBar describing the problem
will be shown. If the connection attempt succeeds, the button turns into a
disconnect button. While the connection attempt is occurring, the button shows
a ProgressRing spinner to indicate an ongoing attempt.

Also add the Connected property to TelloApi.