Implement gamepad flight controls
Add InputController, which processes input from the gamepad (in this case, it
should be a PlayStation DualShock4 controller). This class is capable of feeding
RC control messages directly to the TelloApi instance. In the future, it should
be refactored to feed control inputs to a middleman class for button mappings
and event-based inputs.