Fix misc freezes in Flight module
Fix the process not fully terminating when exiting the Flight window without an
active connection to the drone. Calling TelloApi.StopVideo() without an active
connection will cause it to hang forever while waiting for the drone to respond.
The workaround is simply to check TelloApi.Connected before calling StopVideo().

Fix a similar potential freeze to the one described above when disconnecting
from the drone with the Connect (Disconnect) button.