Convert CDI Position event hook to DispatcherTimer
In ReplaySynchronizer, convert the CDI's
MediaPlayerElement.MediaPlayer.PlaybackSession.PositionChanged event hook, which
updates the current flight state in the CDI to keep it in sync with video
playback, to an independent DispatcherTimer with an interval of 10ms.

This new hook polls the Position field of the MediaPlayerElement, which is more
performant than hooking the PositionChanged event directly, per MS docs. It will
also allow for a higher flight state refresh rate as compared to the ~4/s rate
that the PositionChanged hook runs at.
See
https://learn.microsoft.com/en-us/uwp/api/windows.ui.xaml.controls.mediaplayerelement?view=winrt-22621
for more details.