Use MediaTranscoder instead of ffmpeg
Switch from using ffmpeg to re-encode the mission video after it has been saved
to using Windows.Media.Transcoding.MediaTranscoder. This improves the final
result of the encoding process, as the timestamp hints left in the MediaSamples
while they are captured from the drone's H264 stream are preserved in the
output, whereas with ffmpeg they were lost, leading to desyncs during playback.