Comprehensive UI rework
Convert the MainWindow of DroneNavigator (i.e. the app's landing page) to use a
NavigationView as its primary control. The NavigationView contains two pages:
1. MissionListPage
2. DroneListPage

MissionListPage is the default page, and contains a list of previously flown
missions, which is the same as the previous landing screen. The search bar (not
yet functional) and "Start new mission" button remain unchanged at the bottom of
the list. The margin between list items has been increased to 4. Each list item
now also shows the name and description of the mission. The expander for each
list item also now shows the date and time that the mission started, as well as
the mission duration.

DroneListPage is an entirely new page for managing drones. Drones can be
created, edited, and deleted from this page. DroneListPage is structured
similarly to MisisonListPage; a list of drones is the main focus of the page,
along with a button to create a new drone at the bottom of the page, which
launches a ContentDialog. Each drone in the list has buttons to delete (which
currently violates referential integrity until the data model is updated to
allow for "disabling" of old drones when the user chooses to delete them), or
edit, which is a safe operation. Each list item shows information about the
drone, such as the name, make, model, mission count, and total flight time.