Close MainWindow during Flight
Fixes a regression from 0f5e2d4eb7a4c643517f7422986a8d60fc8c07d9. Prior to that
commit, the MainWindow would close when a new mission was successfully started.
The commit commented out that behavior because there was not an existing way to
get a reference to the MainWindow that was the parent of the NavigationFrame in
the new UI structure.

Fix the aforementioned regression by exposing a reference to the active main
window from DroneNavigator.App, referencing that via the Application.Current
helper, and calling its Close method.

Fix another bug where the main window would not reopen after flight if no video
was encoded. This bug occurred because the callback to reinstantiate the
MainWindow was hooked to the OnMissionVideoEncodingCompleted event handler,
which would not fire when no encoding was required (mainly occurring when no
mission was flown / the drone was never connected to). The fix is to simply add
a new FlightFinished event handler to the Flight window class that is fired
when the window is closing, either after encoding finishes or when no encoding
is required. The regular OnClosed event cannot be hooked since it is intercepted
to trigger encoding.