Fix many crashes on dis/connect and page nav
Fix a variety of crashes relating to object reuse when connecting and
disconnecting from the drone. Fix not being able to connect the drone when
starting the mission before connecting to the drone's WiFi AP. Fix a crash when
closing out of an ongoing mission when running outside of the IDE (likely
related to improper resource freeing).

Move the thread initialization and startup to the Connect button's click
handler, which gets us a clean state to start each new drone connection with.
Also, fully clean up all input threads, networking threads, and network sockets
when disconnecting to avoid doubled resource requisition (caused crashes when
trying to use already-in-use (i.e. not-yet-freed) sockets when connecting to the
drone a second time).
Additional behavior around mission restarts may need reevaluation, such as
the start date and time of a mission being when the Connect button is clicked vs
when the page is opened.

Known bug: The video feed shows only black on consecutive disconnects and
reconnects.