Replace distance stats with cumulative time flown
"Total flown distance" stats are hard to measure accurately for the drone. A
potential replacement would be the maximum distance from the controller that the
drone has flown (which is the Distance field returned in the drone state
packets), but for now, cumulative time flown and average time flown are good
metrics to display.