Fix video-related crashes
Fix a bug where, after the first mission, the application would crash after
being unable to find mission_video.raw, which is now unused.
Fix a bug where the Flight module would crash if the Disconnect button was
clicked before clicking Close, which was due to improper tracking of the mission
video samples variable when cleaning up on Disconnect.