Fix crashes when launching Review
Fix two separate but related crashes that would occur when launching the Review
module.
The first crash was related to missing mission videos; now, if the mission video
does not exist, an error will be displayed when launching Review.
The second was related to missing mission state records in the FlightStates
table; now, if there are no FlightStates for a given mission, an error will be
displayed when launching Review.