Avoid respawning Main window after mission
Change the approach used to block user input to the main module while the Flight
module runs. Instead of closing down the Window entirely, which was causing
some intermittent crashes when run outside of the IDE (likely due to
concurrency issues), simply launch a non-closable ContentDialog that blocks user
input to the Main module's UI while the Flight module is open. Once the Flight
module closes (after the mission is complete and video transcoding finishes),
the ContentDialog will be hidden and the mission list refreshed, allowing the
user to continue normally.