Fix live video stuttering
Fixes an issue where video being streamed live from the drone would stutter.
Only I-frames would be displayed. This was due to the MediaPlayerElement
attempting to skip samples to remain caught up to the real-time stream.
The solution is to disable real-time mode on the MediaPlayerElement.
Unfortunately, this also introduces around 3/4 of a second of latency.
Transcoded video output and subsequent replay synchronization are unaffected.

H/T to Mark Lauter for leaving a comment in his Tello API implementation:
https://github.com/marklauter/TelloAPI-SDK-2.0/blob/master/src/Tello.App.UWP/MainPage.xaml.cs#L122